# frontend/src/api/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "ExportAccepted": {
        "description": "Réponse du chemin asynchrone : la requête HTTP rend la main immédiatement.",
        "properties": {
          "poll_url": {
            "title": "Poll Url",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "default": "queued",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "poll_url"
        ],
        "title": "ExportAccepted",
        "type": "object"
      },
      "ExportStatus": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "state",
          "ready"
        ],
        "title": "ExportStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Éditeur de plan de rénovation 2D → 3D",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/me/export": {
      "get": {
        "description": "Export complet des données du compte, en JSON (RGPD art. 15 et 20).\n\nLe périmètre est exactement celui que les routes de l'API autorisent : les organisations dont\nle compte est membre **accepté**, et ce qu'elles portent. Un export plus large serait une\nfuite entre locataires déguisée en conformité, ce que `tests/test_rgpd.py` vérifie.",
        "operationId": "export_current_account_api_auth_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Current Account Api Auth Me Export Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Export Current Account",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/projects/{project_id}/exports/pdf": {
      "post": {
        "description": "Demande un export PDF en tâche de fond.\n\nLa réponse est immédiate : c'est tout l'intérêt du passage à Celery pour un traitement dont\nla durée croît avec la taille du plan.",
        "operationId": "request_pdf_export_api_projects__project_id__exports_pdf_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Request Pdf Export",
        "tags": [
          "export"
        ]
      }
    },
    "/api/projects/{project_id}/exports/pdf/direct": {
      "get": {
        "description": "Génère et renvoie le PDF dans la requête (chemin synchrone de référence).\n\nConservé délibérément : c'est le point de comparaison qui rend mesurable le gain apporté par\nCelery, comme l'exige `docs/spec-complete.md` §8 (cas 2). `measure=true` ajoute la durée de\ngénération dans un en-tête, pour comparer sans instrumenter le client.",
        "operationId": "export_pdf_synchronously_api_projects__project_id__exports_pdf_direct_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "measure",
            "required": false,
            "schema": {
              "default": false,
              "title": "Measure",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Export Pdf Synchronously",
        "tags": [
          "export"
        ]
      }
    },
    "/api/projects/{project_id}/exports/tasks/{task_id}": {
      "get": {
        "description": "État d'un export.\n\nLa route porte l'identifiant du projet et le vérifie : être authentifié ne suffit pas. Sans\nce cloisonnement, n'importe quel compte connaissant un identifiant de tâche lirait le\ndescriptif de l'export d'autrui — nom de fichier, taille, identifiant de projet.\n\nLe résultat n'expose aucun chemin absolu, et le téléchargement repasse par une route qui\nrevérifie elle aussi la propriété.",
        "operationId": "read_export_status_api_projects__project_id__exports_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Export Status",
        "tags": [
          "export"
        ]
      }
    },
    "/api/projects/{project_id}/exports/{filename}": {
      "get": {
        "description": "Téléchargement d'un export déjà généré.\n\nLe motif du nom de fichier est verrouillé par le routeur, et le chemin est reconstruit à\npartir du répertoire d'export : un `../` n'a aucun moyen d'y entrer (traversée de chemin).\nLa propriété du projet est revérifiée ici, pas seulement à la demande d'export.",
        "operationId": "download_export_api_projects__project_id__exports__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "pattern": "^projet-\\d+-[A-Za-z0-9\\-]+\\.pdf$",
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Download Export",
        "tags": [
          "export"
        ]
      }
    },
...
  }
}